Removed duplcate test
