+++ packages/ilib-loctool-webos-json-resource/test/JSONResourceFile.test.js
@@ -1173,76 +1173,4 @@ describe("jsonresourcefile", function() {
         expect(fs.existsSync(path.join(dir, "fluttermanifest.json"))).toBeFalsy();
         expect(actual).toBe(expected);
     });
-    test("JSONResourceFileWriteManifestWithTargetDir", function() {
-        expect.assertions(5);
-
-        var jsrf = new JSONResourceFile({
-            project: p5,
-            locale: "es-ES"
-        });
-        expect(jsrf).toBeTruthy();
-
-        [
-            p5.getAPI().newResource({
-                type: "string",
-                project: "webosApp",
-                targetLocale: "es-ES",
-                key: "Good Morning!",
-                sourceLocale: "en-KR",
-                source: "Good Morning!",
-                target: "¡Buenos días!"
-            }),
-        ].forEach(function(res) {
-            jsrf.addResource(res);
-        });
-        var expected =
-            '{\n' +
-            '    "Good Morning!": "¡Buenos días!"\n' +
-            '}';
-
-        var actual = jsrf.getContent();
-        diff(actual, expected);
-
-        jsrf.write();
-        var resourceRoot = path.join(p5.target, "resources");
-        jsrf.writeManifest(resourceRoot);
-
-        expect(fs.existsSync(resourceRoot)).toBeTruthy();
-        expect(fs.existsSync(path.join(resourceRoot, "ilibmanifest.json"))).toBeTruthy();
-        expect(jsrf.isDirty()).toBeTruthy();
-        expect(actual).toBe(expected);
-    });
-    test("JSONResourceFileTypeWithTargetDir", function() {
-        expect.assertions(4);
-
-        var jrfile = new JSONResourceFile({
-            project: p5,
-            locale: "es-ES"
-        });
-        var jrftype = new JSONResourceFileType(p5);
-        expect(jrfile).toBeTruthy();
-        expect(jrftype).toBeTruthy();
-
-        [
-            p5.getAPI().newResource({
-                type: "string",
-                project: "webosApp",
-                targetLocale: "es-ES",
-                key: "Good Morning!",
-                sourceLocale: "en-KR",
-                source: "Good Morning!",
-                target: "¡Buenos días!"
-            }),
-        ].forEach(function(res) {
-            jrfile.addResource(res);
-        });
-
-        jrfile.write();
-        jrftype.projectClose();
-
-        var resourceRoot = path.join(p5.target, "resources");
-        expect(fs.existsSync(resourceRoot)).toBeTruthy();
-        expect(fs.existsSync(path.join(resourceRoot, "ilibmanifest.json"))).toBeTruthy();
-    });
-
 });
